refactor vscoderemotesshconfig to be sshconfig
